frontend: Rename alert system
